Refactor events to concern trait
